Rename hosting flow forking related components
